requirements.txt
# ========== 核心依赖 ==========
# Qt 框架（PyQt6）
PyQt6>=6.6.0

# 浏览器自动化（xunika 技术栈）
DrissionPage==4.1.1.2
colorama==0.4.6

# 工具库
cryptography>=41.0.0
pyjwt>=2.8.0
requests>=2.31.0
loguru>=0.7.0
pywin32>=306

# 打包工具
pyinstaller>=6.0.0

